# contracts/openapi/mobile-v1.yaml
openapi: 3.1.0
info:
  title: 404 Mobile API
  version: 0.1.0
  description: >
    Modular monolith baseline contract for the 404 mobile application.
servers:
  - url: /api/v1
...
paths:
...
  /users/me:
    get:
      tags: [User]
      summary: Get my profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: My profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfileResponse'
    patch:
      tags: [User]
      summary: Update my profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfileResponse'
    delete:
      tags: [User]
      summary: Withdraw user
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Withdrawn
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    MyProfileResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        nickname:
          type: string
        profileImageUrl:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
    UpdateProfileRequest:
      type: object
      properties:
        nickname:
          type: string
        profileImageUrl:
          type: string
          nullable: true
...